+++ packages/headless/src/app/reducers.ts
@@ -24,6 +24,7 @@ import {querySuggestReducer} from '../features/query-suggest/query-suggest-slice
 import {queryReducer} from '../features/query/query-slice';
 import {questionAnsweringReducer} from '../features/question-answering/question-answering-slice';
 import {recentResultsReducer} from '../features/recent-results/recent-results-slice';
+import {recentQueriesReducer} from '../features/recent-queries/recent-queries-slice';
 import {recommendationReducer} from '../features/recommendation/recommendation-slice';
 import {redirectionReducer} from '../features/redirection/redirection-slice';
 import {resultPreviewReducer} from '../features/result-preview/result-preview-slice';
@@ -68,6 +69,7 @@ export const triggers = triggerReducer;
 export const questionAnswering = questionAnsweringReducer;
 export const standaloneSearchBoxSet = standaloneSearchBoxSetReducer;
 export const recentResults = recentResultsReducer;
+export const recentQueries = recentQueriesReducer;
 
 export const history = undoable({
   actionTypes: {