+++ packages/headless/src/test/mock-state.ts
@@ -29,6 +29,7 @@ import {getTriggerInitialState} from '../features/triggers/triggers-state';
 import {getQuestionAnsweringInitialState} from '../features/question-answering/question-answering-state';
 import {getStandaloneSearchBoxSetInitialState} from '../features/standalone-search-box-set/standalone-search-box-set-state';
 import {getRecentResultsInitialState} from '../features/recent-results/recent-results-state';
+import {getRecentQueriesInitialState} from '../features/recent-queries/recent-queries-state';
 
 export function createMockState(
   config: Partial<SearchAppState> = {}
@@ -65,6 +66,7 @@ export function createMockState(
     questionAnswering: getQuestionAnsweringInitialState(),
     standaloneSearchBoxSet: getStandaloneSearchBoxSetInitialState(),
     recentResults: getRecentResultsInitialState(),
+    recentQueries: getRecentQueriesInitialState(),
     ...config,
   };
 }