+++ packages/quantic/force-app/main/default/lwc/exampleStandaloneSearchbox/exampleStandaloneSearchbox.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class ExampleStandaloneSearchbox extends LightningElement {}