+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.52.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.51.1...@coveo/quantic@0.52.0) (2021-08-23)
+
+
+### Features
+
+* **quantic:** add tab component ([#1114](https://github.com/coveo/ui-kit/issues/1114)) ([182b5c3](https://github.com/coveo/ui-kit/commit/182b5c32585aeb46cc3f37c0739b7b18ab2a1c3e))
+
+
+
+
+
+## [0.51.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.51.0...@coveo/quantic@0.51.1) (2021-08-20)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+# [0.51.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.50.1...@coveo/quantic@0.51.0) (2021-08-20)
+
+
+### Features
+
+* **quantic:** standalone search box ([#1107](https://github.com/coveo/ui-kit/issues/1107)) ([88b26bb](https://github.com/coveo/ui-kit/commit/88b26bbf3a148cb77ce66cb0b1a7d5c9f77275bc))
+
+
+
+
+
 ## [0.50.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@0.50.0...@coveo/quantic@0.50.1) (2021-08-19)
 
 **Note:** Version bump only for package @coveo/quantic